* corrected queue build rendering (GL_DEPTH_TEST)
* added GetUnitHealth(), GetTeamResources(), and GetCommandQueue() to the lua GUI interface
* enabled loading the ctrlpanel.lua file from the VFS

* added 'wait queued' and 'selfd queued' to the default key bindings
* corrected WAIT behaviour for factories
* enabled queuing of Self Destruction commands
* added a pathetic SELFD cursor

* added 'alwaysDrawQueue' and 'buildBoxesOnShift' to cmdcolors.txt

* added the 'invqueuekey [0|1]' action, and [InvertQueueKey=0] config parameter  
* added mouse chording for immediate mode access when using inverted queue keying
  (this should probably be ditched, i doubt that anyone will find it very useful)


git-svn-id: http://springrts.com/svn/spring/trunk/tools/MapConv@2430 37977431-3df6-0310-b722-df95706aa16b

+++ FeatureCreator.cpp
@@ -4,6 +4,7 @@
 #define treetop 215 /* TODO Shouldn't this be related to the previous define somehow? */
 extern float* heightmap;
 
+#include <math.h>
 #include <iostream> /* cout */
 using namespace std; /* cout */
 
@@ -207,8 +208,8 @@ void CFeatureCreator::PlaceVent(int x, int y, CBitmap * feature, CBitmap * vent,
 			x=bx+rand()*(40)/RAND_MAX-20;
 			y=by+rand()*(40)/RAND_MAX-20;
 #else
-			x=bx+( (float)(rand())*(40.0) )/RAND_MAX-20;
-			y=by+( (float)(rand())*(40.0) )/RAND_MAX-20;
+			x=(int)(bx + ((float)(rand())*(40.0) )/RAND_MAX-20);
+			y=(int)(by + ((float)(rand())*(40.0) )/RAND_MAX-20);
 #endif
 			if(x<5)
 				x=5;
